requirements.txt
openai-whisper
torch
numpy
SpeechRecognition
pyaudio
